Correct minor bug in choosing destination for node.
Before, we always assumed that, once node `B` was chosen, the
destination node `A` would be its parent. However, if `B` was ancestral
to `A`, the nodes would be swapped, such that the new parent of `B`
would be whatever the old parent of `A` was, not `A` itself. I've now
corrected this.

Additionally, I changed the nomenclature from `W_subtree` to `W_nodes`,
reflecting that we don't always move whole subtrees when doing tree
updates. I also changed from `W_parents` to `W_dests` to reflect that
the `A` we sample may not end up as `B`'s parent.